requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
google-generativeai>=0.3.0
chromadb>=0.4.15
pydantic>=2.5.0
python-dotenv>=1.0.0
slowapi>=0.1.9
redis>=5.0.1
tqdm>=4.66.1
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0
streamlit>=1.28.0